Matched GlCutDLL

+++ src/core2/1EC3810.c
@@ -2,7 +2,7 @@
 #pragma GLOBAL_ASM("asm/nonmatchings/core2/1EC3810/func_800E9F20.s")
 
 // Get World Section
-u16 func_800EA05C(void) 
+s32 func_800EA05C(void) 
 {
 	return D_80132DC2;
 }
